fixed list declaration and initialization java-scala classes[ex 7-8]
