--- src/main/java/main.java
@@ -1,10 +1,0 @@
-
-public class main {
-    public main(){
-
-    }
-    public static void main(String[] args) {
-        System.out.println("Hello, World!");
-    }
-
-}
