--- src/main/java/main/Main.java
@@ -1,7 +1,0 @@
-//commento per Linda
-//commento 2
-class Main {
-    public static void main(String[] args) {
-        System.out.println("Hello World!");
-    }
-}
