+++ src/main/scala/Coucou.scala
@@ -5,11 +5,13 @@ object geeks
 {
   def main(args: Array[String])
   {
-    val students = new ArrayList[String]("Linda", "Alberto", "Federico", "Luca")
-
+    val students = new ArrayList[String]()
+    students.add("Linda")
+    students.add("Marco")
+    students.add("Filippo")
     for (student <- students)
     {
-      println("Salut" + student)
+      println("Salut " + student)
     }
   }
 }
